+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from '@mui/material';
 import CssBaseline from '@mui/material/CssBaseline';
 import 'src/translations/i18n';
-import { Layout } from 'src/components/Layout';
+import { Layout } from 'src/components/Layout/Layout';
 import { checkoutTheme } from 'src/style';
 import { CreateRtpPage } from 'src/pages/CreateRtpPage';
