+++ src/pages/CreateRtpPage/index.tsx
@@ -1,0 +1,1 @@
+export { CreateRtpPage } from './CreateRtpPage';
